projects/07-tensorrt-llm-optimization/requirements.txt
# TensorRT-LLM and Core Dependencies
tensorrt-llm>=0.7.0
tensorrt>=8.6.0

# PyTorch and ML Framework
torch>=2.1.0
torchvision
torchaudio

# Transformers and Model Hub
transformers>=4.36.0
accelerate>=0.24.0
datasets>=2.14.0

# Tokenizers
sentencepiece>=0.1.99
tiktoken>=0.5.0
tokenizers>=0.15.0

# ONNX and Model Conversion
onnx>=1.15.0
onnxruntime-gpu>=1.16.0

# Quantization Libraries
optimum>=1.14.0
auto-gptq>=0.5.0
autoawq>=0.1.6

# Data Processing and Utilities
numpy>=1.24.0
pandas>=2.0.0
scipy>=1.11.0
scikit-learn>=1.3.0

# Visualization and Analysis
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.17.0

# Jupyter and Interactive Development
jupyter>=1.0.0
ipywidgets>=8.1.0
tqdm>=4.66.0

# Configuration and Logging
pyyaml>=6.0.0
omegaconf>=2.3.0
wandb>=0.16.0

# Performance Monitoring
psutil>=5.9.0
gpustat>=1.1.0
nvidia-ml-py>=12.535.0

# Development and Testing
pytest>=7.4.0
black>=23.9.0
flake8>=6.1.0
mypy>=1.6.0

# CUDA Utilities (Optional)
cupy-cuda12x>=12.3.0
pynvml>=11.5.0